Cleaned the directory and added some support for non-fits images
